// frontend/src/api/generated/types.gen.ts
export const CharacterClass = {
    /**
     * MERCENARY
     */
    MERCENARY: 0,
    /**
     * EXPLORER
     */
    EXPLORER: 2,
    /**
     * NOBLE
     */
    NOBLE: 3,
    /**
     * SAINT
     */
    SAINT: 4,
    /**
     * GLADIATOR
     */
    GLADIATOR: 5,
    /**
     * GUARDIAN_SWORDSMAN
     */
    GUARDIAN_SWORDSMAN: 6,
    /**
     * EXCAVATOR
     */
    EXCAVATOR: 7,
    /**
     * SNIPER
     */
    SNIPER: 8,
    /**
     * COURT_MAGICIAN
     */
    COURT_MAGICIAN: 9,
    /**
     * MAGIC_KNIGHT
     */
    MAGIC_KNIGHT: 10,
    /**
     * PRIEST
     */
    PRIEST: 11,
    /**
     * SHAMAN
     */
    SHAMAN: 12
} as const;

/**
 * CharacterClass
 */
export type CharacterClass = typeof CharacterClass[keyof typeof CharacterClass];

/**
 * HTTPValidationError
 */
export type HttpValidationError = {
    /**
     * Detail
     */
    detail?: Array<ValidationError>;
};

/**
 * MatchParticipant
 */
export type MatchParticipant = {
    /**
     * Player Id
     */
    player_id: number;
    /**
     * Name
     */
    name: string;
    /**
     * Level Land
     */
    level_land: number;
    /**
     * Level Sea
     */
    level_sea: number;
    character_class: CharacterClass;
    /**
     * Is Female
     */
    is_female: boolean;
    /**
     * Elo Before
     */
    elo_before: number | null;
    /**
     * Elo Change
     */
    elo_change: number | null;
};

/**
 * MatchRead
 */
export type MatchRead = {
    /**
     * Id
     */
    id: number;
    /**
     * Round Id
     */
    round_id: number;
    team_1: MatchTeam;
    team_2: MatchTeam;
    match_result: MatchResult;
};

/**
 * MatchResult
 */
export const MatchResult = {
    /**
     * UNSET
     */
    UNSET: 1,
    /**
     * CANCELED
     */
    CANCELED: 2,
    /**
     * WIN_TEAM_1
     */
    WIN_TEAM_1: 3,
    /**
     * WIN_TEAM_2
     */
    WIN_TEAM_2: 4,
    /**
     * DRAW
     */
    DRAW: 5
} as const;

/**
 * MatchResult
 */
export type MatchResult = typeof MatchResult[keyof typeof MatchResult];

/**
 * MatchTeam
 */
export type MatchTeam = {
    /**
     * Members
     */
    members: Array<MatchParticipant>;
};